# docs/architecture/api_specifications.yaml
openapi: 3.0.3
info:
  title: NumerAI API
  description: |
    AI-Powered Numerology Platform API
    
    ## Authentication
    Most endpoints require JWT authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    - Free tier: 20 requests/minute
    - Premium tier: 100 requests/minute
    
    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets
    
    ## Error Handling
    All errors follow a consistent format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message",
        "details": {}
      }
    }
    ```
  version: 1.0.0
  contact:
    name: NumerAI API Support
    email: api@numerai.app
  license:
    name: Proprietary
servers:
  - url: https://api.numerai.app/api/v1
    description: Production server
  - url: https://staging-api.numerai.app/api/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /auth/refresh-token:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                      refresh_token:
                        type: string
                        description: New refresh token (if rotation enabled)
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: object
...